Avoid inconvertible types warning at the `assert composite*.id == [1,2]` as left hand side is `ArrayList<Long>` while right hand side is `ArrayList<Integer>`
(cherry picked from commit 5dec8691ff0c4da4c2cb8c135cf93775092b9ffc)

+++ src/spec/test/OperatorsTest.groovy
@@ -421,7 +421,7 @@ assert null*.make == null                             // <3>
         assertScript '''
 // tag::spreaddot_iterable[]
 class Component {
-    Long id
+    Integer id
     String name
 }
 class CompositeObject implements Iterable<Component> {
